Allow setting a description for boot environments.
This commit implements a useful feature of beadm from Solaris/illumos:
human-friendly "descriptions" for boot environments.

As in Solaris/illumos, this description can be set with zectl create -d.
Unlike Solaris/illumos, though, I've also taken the liberty of adding it
to the zectl list output. For example:

    $ zectl create -d "Testing COSMIC alpha" pop-cosmic

    $ zectl list
    Name        Active  Mountpoint  Space  Creation          Description
    pop-cosmic          -           8K     2024-02-01 21:12  Testing COSMIC alpha
    arch-demo           -           8K     2024-01-27 10:06  -
    chimera             -           2.90G  2024-01-23 15:12  Chimera Linux
    pop         NR      /           36.1G  2024-01-24 11:52  Pop!_OS 22.04 LTS
    arch                -           2.70G  2023-03-25 21:57  Arch Linux

On the library side there is a new public libze_set_description()
function that can be used by other frontends, and a "description"
property will appear in the result of libze_list() if present.

Internally, the description is just a well-known namespaced property, as
with the existing bootloader properties.

Under Solaris/illumos this description would also be used to populate
the boot menu, and we could do the same thing via bootloader plugins in
the future, too.

Signed-off-by: Aaron Jacobs <atheriel@gmail.com>